+++ udata/tests/api/test_topics_api.py
@@ -31,6 +31,20 @@ class TopicsAPITest(APITestCase):
         self.assert200(response)
         self.assertEqual(len(response.json['data']), 1)
         self.assertEqual(response.json['data'][0]['id'], str(tag_topic.id))
+        datasets = response.json['data'][0]['datasets']
+        self.assertEqual(len(datasets), 3)
+        for dataset, expected in zip(datasets, [d.fetch() for d in tag_topic.datasets]):
+            self.assertEqual(dataset['id'], str(expected.id))
+            self.assertEqual(dataset['title'], str(expected.title))
+            self.assertIsNotNone(dataset['page'])
+            self.assertIsNotNone(dataset['uri'])
+        reuses = response.json['data'][0]['reuses']
+        for reuse, expected in zip(reuses, [r.fetch() for r in tag_topic.reuses]):
+            self.assertEqual(reuse['id'], str(expected.id))
+            self.assertEqual(reuse['title'], str(expected.title))
+            self.assertIsNotNone(reuse['page'])
+            self.assertIsNotNone(reuse['uri'])
+        self.assertEqual(len(reuses), 3)
 
     def test_topic_api_get(self):
         '''It should fetch a topic from the API'''
@@ -39,8 +53,17 @@ class TopicsAPITest(APITestCase):
         self.assert200(response)
 
         data = response.json
-        for dataset, expected in zip(data['datasets'], topic.datasets):
+        for dataset, expected in zip(data['datasets'], [d.fetch() for d in topic.datasets]):
             self.assertEqual(dataset['id'], str(expected.id))
+            self.assertEqual(dataset['title'], str(expected.title))
+            self.assertIsNotNone(dataset['page'])
+            self.assertIsNotNone(dataset['uri'])
+
+        for reuse, expected in zip(data['reuses'], [r.fetch() for r in topic.reuses]):
+            self.assertEqual(reuse['id'], str(expected.id))
+            self.assertEqual(reuse['title'], str(expected.title))
+            self.assertIsNotNone(reuse['page'])
+            self.assertIsNotNone(reuse['uri'])
 
     def test_topic_api_create(self):
         '''It should create a topic from the API'''